separate price tolerance to a separate hook
