+++ dapp/src/hooks/usePriceTolerance.js
@@ -1,0 +1,32 @@
+import React, { useState, useEffect } from 'react'
+
+const usePriceTolerance = (mode) => {
+  const [priceToleranceValue, setPriceTolerFct] = useState(null)
+  const priceToleranceLocalStorageKey = `selected_price_tolerance_${mode}`
+
+  const dropdownToleranceOptions = [0.25, 0.5, 1]
+
+  // store price tolerance value for future default setting
+  useEffect(() => {
+    // default price tolerance value
+    let priceTolerance = 0.5
+    let localStorageValue = localStorage.getItem(priceToleranceLocalStorageKey)
+    if (localStorageValue) {
+      priceTolerance = parseFloat(localStorageValue)
+    }
+    setPriceToleranceValue(priceTolerance)
+  }, [])
+
+  const setPriceToleranceValue = (value) => {
+    setPriceTolerFct(value)
+    localStorage.setItem(priceToleranceLocalStorageKey, value)
+  }
+
+  return {
+    setPriceToleranceValue,
+    priceToleranceValue,
+    dropdownToleranceOptions,
+  }
+}
+
+export default usePriceTolerance
